basic changes to support copy user/pass from PwEntryView
