+++ app/src/main/java/com/keepassdroid/EntryActivity.java
@@ -20,31 +20,20 @@
  */
 package com.keepassdroid;
 
-import java.text.DateFormat;
-import java.util.Date;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.UUID;
-
 import android.app.Activity;
-import android.app.AlertDialog;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.ActivityNotFoundException;
 import android.content.BroadcastReceiver;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
 import android.preference.PreferenceManager;
-import android.text.SpannableString;
-import android.text.method.LinkMovementMethod;
 import android.text.method.PasswordTransformationMethod;
-import android.text.util.Linkify;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
@@ -66,6 +55,12 @@ import com.keepassdroid.intents.Intents;
 import com.keepassdroid.utils.EmptyUtils;
 import com.keepassdroid.utils.Types;
 import com.keepassdroid.utils.Util;
+
+import java.text.DateFormat;
+import java.util.Date;
+import java.util.Timer;
+import java.util.TimerTask;
+import java.util.UUID;
 
 public class EntryActivity extends LockCloseHideActivity {
 	public static final String KEY_ENTRY = "entry";
@@ -307,7 +302,7 @@ public class EntryActivity extends LockCloseHideActivity {
 		} else {
 			togglePassword.setTitle(R.string.menu_showpass);
 		}
-		
+
 		MenuItem gotoUrl = menu.findItem(R.id.menu_goto_url);
 		MenuItem copyUser = menu.findItem(R.id.menu_copy_user);
 		MenuItem copyPass = menu.findItem(R.id.menu_copy_pass);
@@ -410,15 +405,15 @@ public class EntryActivity extends LockCloseHideActivity {
 		try {
 			Util.copyToClipboard(this, text);
 		} catch (SamsungClipboardException e) {
-			showSamsungDialog();
+			Util.showSamsungDialog(this);
 			return;
 		}
-		
+
 		SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
 		String sClipClear = prefs.getString(getString(R.string.clipboard_timeout_key), getString(R.string.clipboard_timeout_default));
-		
+
 		long clipClearTime = Long.parseLong(sClipClear);
-		
+
 		if ( clipClearTime > 0 ) {
 			mTimer.schedule(new ClearClipboardTask(this, text), clipClearTime);
 		}
@@ -452,27 +447,5 @@ public class EntryActivity extends LockCloseHideActivity {
 				}
 			}
 		}
-	}
-	
-	private void showSamsungDialog() {
-		String text = getString(R.string.clipboard_error).concat(System.getProperty("line.separator")).concat(getString(R.string.clipboard_error_url));
-		SpannableString s = new SpannableString(text);
-		TextView tv = new TextView(this);
-		tv.setText(s);
-		tv.setAutoLinkMask(RESULT_OK);
-		tv.setMovementMethod(LinkMovementMethod.getInstance());
-		Linkify.addLinks(s, Linkify.WEB_URLS);
-		
-		AlertDialog.Builder builder = new AlertDialog.Builder(this);
-		builder.setTitle(R.string.clipboard_error_title)
-			.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-				@Override
-				public void onClick(DialogInterface dialog, int which) {
-					dialog.dismiss();
-				}
-			})
-			.setView(tv)
-			.show();
-		
 	}
 }
