+++ app/src/main/res/values/strings.xml
@@ -133,7 +133,7 @@
     <string name="menu_about">About</string>
     <string name="menu_change_key">Change Master Key</string>
     <string name="menu_copy_pass">Copy Password</string>
-    <string name="menu_copy_user">Copy User</string>
+    <string name="menu_copy_user">Copy Username</string>
     <string name="menu_create">Create</string>
     <string name="menu_app_settings">Settings</string>
     <string name="menu_db_settings">Database settings</string>
